+++ src/features/settings/general/other/backup/helpers.ts
@@ -1,7 +1,7 @@
 import { pickBy, without } from "lodash";
 import { db } from "../../../../../services/db";
 import { getAllObjectValuesDeep } from "../../../../../helpers/object";
-import { LOCALSTORAGE_KEYS, get, set } from "../../../storage";
+import { LOCALSTORAGE_KEYS, get, set } from "../../../syncStorage";
 
 const BASE_BACKUP_JSON = {
   voyagerBackupVersion: 1,
